don't need whole script to fix concurrent example
